AsyncBackend/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioice==0.7.6
aiortc==1.3.2
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
deepface==0.0.75
dnspython==2.2.1
filelock==3.7.0
fire==0.4.0
Flask==2.1.2
flatbuffers==1.12
frozenlist==1.3.0
gast==0.4.0
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-crc32c==1.3.0
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mtcnn==0.1.1
multidict==6.0.2
netifaces==0.11.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==9.0.4
pylibsrtp==0.7.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
